+++ cmdVersion/questionFactory/AnimeComparatorFactory.java
@@ -1,10 +1,11 @@
-package cmdVersion;
-
+package cmdVersion.questionFactory;
 import connection.Anime;
-
 import java.util.Comparator;
 
+
+
 public class AnimeComparatorFactory {
+    final public String[] comparatorType = {"ratingRank"};
 
     public static Comparator<Anime> getComparator(String type){
         Comparator<Anime> ratingRankComp = (a1,a2) -> a2.get_rating_rank()-a1.get_rating_rank(); // Positive when a1 has better rank than a2
