+++ cmdVersion/questionFactory/AnimeConstantFactory.java
@@ -1,0 +1,4 @@
+package cmdVersion.questionFactory;
+
+public class AnimeConstantFactory {
+}
