+++ cmdVersion/questionFactory/AnimePromptFactory.java
@@ -1,0 +1,5 @@
+package cmdVersion.questionFactory;
+
+public class AnimePromptFactory {
+
+}
